+++ .gitignore
@@ -3,3 +3,8 @@ node_modules
 dist/*.*
 semantic
 semantic.json
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
